[MODIFIED]- product module, restricted user access

+++ src/app/modules/product/product.service.ts
@@ -7,10 +7,11 @@ import Category from '../category/category.model';
 import PowerSource from '../powerSource/powerSource.model';
 import Connectivity from '../connectivity/connectivity.model';
 import Tag from '../tag/tag.model';
-import OperatingSystem from '../operatingSystem/operatingSystem.model';
 import QueryBuilder from '../../builder/QueryBuilder';
 import { ProductSearchableFields } from './product.constant';
 import { Types } from 'mongoose';
+import User from '../user/user.model';
+import OperatingSystem from '../operatingSystem/operatingSystem.model';
 
 // create product
 const createProduct = async (payload: TProduct, userId: Types.ObjectId) => {
@@ -93,13 +94,26 @@ const createProduct = async (payload: TProduct, userId: Types.ObjectId) => {
 };
 
 // get all product
-const getAllProduct = async (query: Record<string, unknown>) => {
+const getAllProduct = async (
+  query: Record<string, unknown>,
+  userEmail: string,
+) => {
+  // check is user exist
+  const isUserExist = await User.findOne({ email: userEmail });
+  if (!userEmail) {
+    throw new AppError(StatusCodes.UNAUTHORIZED, 'Account does not exist.');
+  }
+
   const queryObj = { ...query };
   delete queryObj['tags'];
   const tags = (query?.tags as string)?.split(',');
 
   if (query.tags) {
     queryObj['tags'] = { $in: tags };
+  }
+
+  if (isUserExist?.role === 'user') {
+    queryObj['createdBy'] = isUserExist._id;
   }
 
   // product query model
@@ -127,7 +141,13 @@ const getAllProduct = async (query: Record<string, unknown>) => {
 };
 
 // get single product by id
-const getSingleProductById = async (id: string) => {
+const getSingleProductById = async (id: string, userEmail: string) => {
+  // check is user exist
+  const isUserExist = await User.findOne({ email: userEmail });
+  if (!isUserExist) {
+    throw new AppError(StatusCodes.UNAUTHORIZED, 'Account does not founded.');
+  }
+
   const result = await Product.findById(id).populate({
     path: 'tags',
     select: '-__v',
@@ -138,15 +158,32 @@ const getSingleProductById = async (id: string) => {
       'Product not founded by this ID.',
     );
   }
+
+  if (isUserExist.role === 'user' && result.createdBy !== isUserExist._id) {
+    throw new AppError(StatusCodes.FORBIDDEN, 'Access forbidden.');
+  }
   return result;
 };
 
 // delete product by Id
-const deleteProductById = async (productId: string) => {
+const deleteProductById = async (productId: string, userEmail: string) => {
+  // check is user exist
+  const isUserExist = await User.findOne({ email: userEmail });
+  if (!isUserExist) {
+    throw new AppError(StatusCodes.UNAUTHORIZED, 'Account does not founded.');
+  }
+
   // check is product exist
   const isProductExist = await Product.findById(productId);
   if (!isProductExist) {
     throw new AppError(StatusCodes.NOT_FOUND, 'Product not founded.');
+  }
+
+  if (
+    isUserExist.role === 'user' &&
+    isProductExist.createdBy !== isUserExist._id
+  ) {
+    throw new AppError(StatusCodes.FORBIDDEN, 'Access forbidden.');
   }
 
   const result = await Product.findByIdAndDelete(productId);
@@ -160,12 +197,30 @@ const deleteProductById = async (productId: string) => {
 };
 
 //update single product by Id
-const updateProductById = async (id: string, payload: Partial<TProduct>) => {
+const updateProductById = async (
+  id: string,
+  payload: Partial<TProduct>,
+  userEmail: string,
+) => {
+  // check is user exist
+  const isUserExist = await User.findOne({ email: userEmail });
+  if (!isUserExist) {
+    throw new AppError(StatusCodes.UNAUTHORIZED, 'Account does not founded.');
+  }
+
   // check if product is exist
   const isProductExist = await Product.findById(id);
   if (!isProductExist) {
     throw new AppError(StatusCodes.NOT_FOUND, 'Product not founded.');
   }
+
+  if (
+    isUserExist.role === 'user' &&
+    isProductExist.createdBy !== isUserExist._id
+  ) {
+    throw new AppError(StatusCodes.FORBIDDEN, 'Access forbidden.');
+  }
+
   if (payload.quantity && payload.quantity > 0) {
     payload.status = 'in-stock';
   } else {
